requirements.txt
scrapy
google-api-python-client
openai
sqlalchemy
requests
python-dotenv
selenium>=4.10.0
pillow>=9.5.0
dnspython>=2.7.0
